feat: add state meta and carousel navigation
