+++ data/stateMeta.ts
@@ -1,3 +1,27 @@
+export const stateMeta: Record<string, { contact: string; role?: string; slug: string; title: string; subtitle: string }> = {
+  niger: {
+    slug: "niger",
+    title: "Niger State",
+    subtitle: "The Power State",
+    contact: "Dr. Matthew Ahmed",
+    role: "Permanent Secretary, Ministry of Agriculture & Food Security",
+  },
+  kwara: {
+    slug: "kwara",
+    title: "Kwara State",
+    subtitle: "The State of Harmony",
+    contact: "Hon. Alabi Afeez",
+    role: "Commissioner of Agriculture",
+  },
+  plateau: {
+    slug: "plateau",
+    title: "Plateau State",
+    subtitle: "The Home of Peace and Tourism",
+    contact: "Mr. Yakubu Nuhu",
+    role: "Special Adviser to the Governor on Carbon",
+  },
+};
+
 export const stateTitles: Record<string, string> = {
   niger: "Niger — Rice & Forestry MRV",
   kwara: "Kwara — Renewable + Agro pilots",
